refactor: Move controller annotation to Api interface - #117
