+++ src/main/java/talkPick/domain/admin/application/AdminQueryService.java
@@ -1,4 +1,4 @@
-package talkPick.domain.admin.application.query;
+package talkPick.domain.admin.application;
 
 import org.springframework.transaction.annotation.Transactional;
