+++ src/main/java/talkPick/domain/admin/application/AuthQueryService.java
@@ -1,0 +1,4 @@
+package talkPick.domain.admin.application;
+
+public class AuthQueryService {
+}
