+++ src/main/java/talkPick/domain/admin/domain/type/Role.java
@@ -1,4 +1,4 @@
-package talkPick.domain.auth;
+package talkPick.domain.admin.domain.type;
 
 public enum Role {
     ADMIN, DEV, MEMBER
