+++ src/main/java/talkPick/domain/admin/validator/ValidationMessages.java
@@ -1,4 +1,4 @@
-package talkPick.domain.admin.application.validator;
+package talkPick.domain.admin.validator;
 
 public class ValidationMessages {
     public static final String PASSWORD_VALIDATION = "password must be at least 8 characters, including 1 uppercase, lowercase, number, and special character";
