--- src/main/java/talkPick/domain/auth/application/query/AuthQueryService.java
@@ -1,4 +1,0 @@
-package talkPick.domain.auth.application.query;
-
-public class AuthQueryService {
-}
