+++ src/main/java/talkPick/domain/member/exception/MemberAlreadyExistsexception.java
@@ -1,4 +1,4 @@
-package talkPick.batch.topic.exception;
+package talkPick.domain.member.exception;
 
 public class MemberAlreadyExistsexception extends RuntimeException {
     public MemberAlreadyExistsexception(String message) {
