+++ src/main/java/talkPick/global/security/model/WhiteList.java
@@ -1,0 +1,4 @@
+package talkPick.global.security.model;
+
+public class WhiteList {
+}
